Added AutoFilter Features
Added the ability for users to add, clear, and delete data filters. Filtering options have been limited for the time being compared to what is capable by the Excel Interop. Updated command item names to try and make more sense. This was more for an ease of user understanding than a functionality change.